dev: move graphql mutations to main api file
